Correct Powershell Env setting
In Powershell env vars are set via 

```powershell
> $Env:DEFMT_LOG = "trace" 
```

Trying to run the example, as formatted prior to this change:

```powershell
> Env: DEFMT_LOG = "trace"  
Env: : The term 'Env:' is not recognized as the name of a cmdlet, function, script file, or operable program. 
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
At line:1 char:1
+ Env: DEFMT_LOG = "trace"
+ ~~~~
    + CategoryInfo          : ObjectNotFound: (Env::String) [], CommandNotFoundException
    + FullyQualifiedErrorId : CommandNotFoundException
``` 

This is documented in https://docs.microsoft.com/en-us/powershell/module/microsoft.powershell.core/about/about_environment_variables?view=powershell-7.2.